+++ src/domain/@shared/event/event.dispatcher.interface.ts
