+++ src/domain/@shared/event/event.handler.interface.ts
