+++ src/domain/@shared/event/event.interface.ts
