+++ src/domain/@shared/repository/repository.interface.ts
