+++ src/domain/checkout/repository/order.repository.interface.ts
@@ -1,5 +1,5 @@
 import Order from "../entity/order";
-import RepositoryInterface from "./repository.interface";
+import RepositoryInterface from "../../@shared/repository/repository.interface";
 
 export default
   interface OrderRepositoryInterface
