+++ src/domain/checkout/service/order.service.spec.ts
@@ -1,4 +1,4 @@
-import Customer from "../entity/customer";
+import Customer from "../../customer/entity/customer";
 import Order from "../entity/order"
 import OrderItem from "../entity/order_item";
 import OrderService from "./order.service";
