+++ src/domain/customer/event/customerAddressChanged.event.ts
@@ -1,4 +1,4 @@
-import EventInterface from "../@shared/event.interface";
+import EventInterface from "../../@shared/event/event.interface";
 
 export default class CustomerAddressChangedEvent implements EventInterface {
   dataTimeOcurred: Date;
