+++ src/domain/customer/event/handler/enviaConsoleLog.handler.ts
@@ -1,5 +1,5 @@
-import EventHandlerInterface from "../../@shared/event.handler.interface";
-import EventInterface from "../../@shared/event.interface";
+import EventHandlerInterface from "../../../@shared/event/event.handler.interface";
+import EventInterface from "../../../@shared/event/event.interface";
 import CustomerCreatedEvent from "../customerCreated.event";
 
 export default
