+++ src/domain/customer/repository/customer.repository.interface.ts
@@ -1,5 +1,5 @@
 import Customer from "../entity/customer";
-import RepositoryInterface from "./repository.interface";
+import RepositoryInterface from "../../@shared/repository/repository.interface";
 
 export default
   interface CustomerRepositoryInterface
