+++ src/domain/product/entity/product.spec.ts
