+++ src/domain/product/entity/product.ts
