+++ src/infrastructure/mocks/customer.repository.mock.ts
@@ -1,7 +1,5 @@
-import Address from "../../domain/entity/address";
-import Customer from "../../domain/entity/customer";
-import CustomerRepositoryInterface from "../../domain/repository/customer.repository.interface";
-import CustomerModel from "../db/sequelize/model/customer.model";
+import Customer from "../../domain/customer/entity/customer";
+import CustomerRepositoryInterface from "../../domain/customer/repository/customer.repository.interface";
 
 export default
   class CustomerRepositoryMock
